game/requirements.txt
tornado~=6.2
python-dotenv